read_nefit_data/requirements.txt
python-dateutil==2.8.1
psycopg2==2.8.4
requests
Flask==1.1.1
Flask-Cors==3.0.8